+++ Crawler/Page.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Crawler
 {
@@ -15,6 +16,18 @@ namespace Crawler
             Images = new List<string>();
             InternalLinks = new List<string>();
             ExternalLinks = new List<string>();
+        }
+
+        public void OrderLists()
+        {
+            this.Images = OrderList(this.Images);
+            this.InternalLinks = OrderList(this.InternalLinks);
+            this.ExternalLinks = OrderList(this.ExternalLinks);
+        }
+
+        private IList<T> OrderList<T>(IList<T> list)
+        {
+            return list.Distinct().OrderBy(item => item).ToList();
         }
     }
 }